refactor: shift app module into separate folder
